+++ front/src/app/app.component.ts
@@ -1,13 +1,13 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterOutlet } from '@angular/router';
-import { K1Component } from './k1/k1.component';
-import { K2Component } from './k2/k2.component';
-import { K3Component } from './k3/k3.component';
-import { TaskComponent } from './task/task.component';
-import { ScenarioComponent } from "./scenario/scenario.component";
-import { BoardComponent } from "./board/board.component";
-import { LeftMenuComponent } from "./left-menu/left-menu.component";
+import { K1Component } from './dev/k1/k1.component';
+import { K2Component } from './dev/k2/k2.component';
+import { K3Component } from './dev/k3/k3.component';
+import { TaskComponent } from './kanban/task/task.component';
+import { ScenarioComponent } from "./kanban/scenario/scenario.component";
+import { BoardComponent } from "./kanban/board/board.component";
+import { LeftMenuComponent } from "./common/left-menu/left-menu.component";
 
 @Component({
     selector: 'app-root',
