+++ front/src/app/common/icon/icon.component.html
