+++ front/src/app/common/icon/icon.component.spec.ts
