+++ front/src/app/common/left-menu/left-menu.component.html
