+++ front/src/app/dev/k1/k1.component.ts
