+++ front/src/app/dev/k2/k2.component.css
