+++ front/src/app/dev/k2/k2.component.spec.ts
