+++ front/src/app/kanban/board/board.component.html
@@ -7,6 +7,7 @@
       <div class="col-sm m-3" style =" border: solid 1px black; background-color: #f0f0f0; border-radius: 10px;">
         In Progress
         <app-scenario [scenario] = "scen2"></app-scenario>
+        <app-scenario [scenario] = "scen"></app-scenario>
       </div>
       <div class="col-sm m-3" style ="border: solid 1px black; background-color: #f0f0f0; border-radius: 10px;">
         Done
