+++ front/src/app/kanban/scenario/scenario.component.html
