+++ front/src/app/kanban/scenario/scenario.component.spec.ts
