+++ front/src/app/kanban/task/task.component.css
