+++ front/src/app/kanban/task/task.component.html
@@ -1,5 +1,5 @@
 <div class="task">
-    <div class="card mb-3">
+    <div class="card">
         <div class="card-header">
             <div class="form-check">
                 <input class="form-check-input" type="checkbox" id="task-checkbox" />
